添加接清单模块下查询功能

+++ src/main/java/com/giantlizardcloud/merchant/controller/SettleController.java
@@ -31,7 +31,7 @@ public class SettleController {
     }
 
     @PutMapping("/{id}")
-    public JSONResult invalidSettle(@PathVariable("id") Integer sid){
+    public JSONResult invalidSettle(@PathVariable("id") Long sid){
         settleService.invalidSettle(sid);
         return JSONResult.ok();
     }
